Potential Evapotranspiration models

ET0_eq : Equilibrium evaporation, slope / (slope + gamma) * Rn

ET0_Penman48 : Penman 1948 equation simplified by Shuttleworth 1993

ET0_Monteith65: Penman-Monteith 1965

ET0_PT72 : Priestley Taylor 1972, ET0_PT72 = ET0_eq * 1.26

ET0_FAO98 : Penman-Monteith reference crop evapotranspiration, FAO56

ET0_eq(Rn, Tair, Pa = atm, ...)

ET0_Penman48(Rn, Tair, Pa = atm, D, wind, z.wind = 2)

ET0_Monteith65(
  Rn,
  Tair,
  Pa = atm,
  D,
  wind,
  z.wind = 2,
  rs = 70,
  rH = NULL,
  ...
)

ET0_PT72(Rn, Tair, Pa = atm, alpha = 1.26, ...)

ET0_FAO98(Rn, Tair, Pa = atm, D, wind, z.wind = 2, tall_crop = FALSE)

Arguments

Rn

land surface net radiation, W m-2

Tair

2m air temperature (degC)

Pa

surface air pressure (kPa)

D

vapor pressure deficit (kPa)

wind

wind speed at the height of z.wind

z.wind

Height where measure the wind speed [m]. Default 10m.

alpha

albedo [0-1]

tall_crop

tall crop or not, see FAO1998 for details

Value

lambda: Latent heat of vaporization, about [2.5 MJ kg-1].

slope : The slope of the saturation vapour pressure curve at certain air temperature Tair, [kPa degC-1].

gamma : Psychrometric constant ([kPa degC-1]), Cp*Pa/(epsilon*lambda)

Eeq : Equilibrium (radiative) evaporation, mm

Evp : Aerodynamic evaporation, mm

ET0 : Potential evapotranspiration, mm

References

Penman equation. Wikipedia 2021. https://en.wikipedia.org/w/index.php?title=Penman_equation

Shuttleworth, W. J. (1993) Evaporation. In: Handbook of Hydrology (ed. by D. Maidment). McGraw-Hill, New York.

Allen, R. G., & Luis S. Pereira. (1998). Crop evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56. European Journal of Agronomy, 34(3), 144-152. doi:10.1016/j.eja.2010.12.001

Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture Organization of the United Nations, 56(97), e156.

Chapter 2 - FAO Penman-Monteith equation, https://www.fao.org/3/x0490E/x0490e06.htm

Examples

et0_pt = ET0_PT72(250, 25, D = 1)
et0_penman48 = ET0_Penman48(250, 25, D = 1, wind = 2)
et0_fao98 = ET0_FAO98(250, 25, D = 1, wind = 2)
ET0_Monteith65(250, 25, D = 1, wind = 2)
#>     lambda     slope      gamma      Eeq      Evp      ET0
#>      <num>     <num>      <num>    <num>    <num>    <num>
#> 1: 2.44175 0.1886818 0.06758237 5.530013 1.342272 6.872284