Budyko Curve

Budyko Curve

budyko(PET.P, par = 2, method = "Fu1981")

ET_budyko(PET, P, par = 2, method = "Fu1981", ...)

budyko_goal(AET.P, PET.P, par = 2, method = "Fu1981", ...)

budyko_fit(AET.P, PET.P, par = 2, method = "Fu1981", ...)

Arguments

PET.P

PET divide P,

PET: potential evapotranspiration, can be calculated by ET0_Penman48().

P: Precipitation

AET.P: AET / P

PET.P: PET / P

par

scalar value, the parameter of Budyko curve

method

one of c("budyko", "Fu1981", "Zhang2001", "Yang2008", "Pike1964"), or all.

PET

potential evapotranspiration (mm)

P

precipitation (mm)

...

ignored

AET.P

AET / P

References

Budyko, M. I., Climate and Life, 508 pp., Academic, San Diego, Calif., 1974.

Fu, Baw-Puh., 1981. On the Calculation of the Evaporation from Land Surface. Chinese Journal of Atmospheric Sciences, 5(1), 23-31.

Zhang, L., Dawes, W. R., & Walker, G. R. (2001). Response of mean annual evapotranspiration to vegetation changes at catchment scale. Water Resources Research, 37(3), 701–708. doi:10.1029/2000WR900325

Yang, H., Yang, D., Lei, Z., & Sun, F. (2008). New analytical derivation of the mean annual water-energy balance equation. Water Resources Research. doi:10.1029/2007WR006135