Net outgoing longwave radiation.

Net outgoing longwave radiation.

cal_Rln(tmax, tmin, ea, Rsi = NULL, Rsi_o = NULL, cld = NULL)

cal_Rln_yang2019(
  Ts,
  Rsi,
  Rsi_toa,
  lat = 30,
  emiss = 0.96,
  n1 = 2.52,
  n2 = 2.37,
  n3 = 0.035
)

Arguments

tmax

Daily maximum air temperature at 2m height [deg Celsius].

tmin

Daily minimum air temperature at 2m height [deg Celsius].

ea

Actual vapor pressure [kPa]. Can be estimated by maximum or minimum air temperature and mean relative humidity.

Rsi

Surface incoming short-wave radiation (Rsi)

Rsi_o

Clear sky incoming shortwave radiation, i. e. extraterrestrial radiation multiply by clear sky transmissivity (i. e. a + b, a and b are coefficients of Angstrom formula. Normally 0.75) [MJ m-2 day-1].

cld

Cloud cover [fraction]. If provided it would be directly used to calculate net outgoing longwave radiation than Rso.

Ts

land surface temperature

Rsi_toa

incoming short-wave radiation at the top of the atmosphere

lat

latitude (in deg)

emiss

Emissivity

n1, n2, n3

parameter for delta_T, see Yang 2019 for details

Value

Net outgoing longwave radiation [MJ m-2 day-1]

References

Yang, Y., & Roderick, M. L. (2019). Radiation, surface temperature and evaporation over wet surfaces. Quarterly Journal of the Royal Meteorological Society, 145(720), 1118–1129. https://doi.org/10.1002/qj.3481

Tu, Z., & Yang, Y. (2022). On the Estimation of Potential Evaporation Under Wet and Dry Conditions. Water Resources Research, 58(4). https://doi.org/10.1029/2021wr031486