Surface net radiation

Surface net radiation

cal_Rn(
  lat,
  J,
  Tmin,
  Tmax,
  ea = NULL,
  RH,
  ssd = NULL,
  cld = NULL,
  Rsi = NULL,
  albedo = 0.23,
  Z = 0,
  ...
)

Arguments

lat

float, latitude

J

integer, day of year

Tmin, Tmax

numeric vector, min and max 2m-air temperature (degC)

ea

Actual vapor pressure (kPa)

RH

Relative humidity (%). If ea provided, RH will be ignored.

ssd

numeric vector, sun shine duration (hour)

cld

(optional) cloud coverage (0-1). At least one of cld and ssd should be provided. If cld is not null, ssd will be ignored.

Rsi

(optional) Surface downward shortwave radiation (MJ m-2 d-1).

If not provided, Rsi will be calculated by (as + bs * nN) * Rsi_toa.

If provided, ssd and cld will be ignored.

albedo

(optional), Rsn = (1 - albedo) Rsi

Z

(optional) elevation (m), for the calculation of Rsi_o

...

ignored

Value

radiation in [MJ d-1]

Rn : Surface net radiation

Rln : Surface outward net longwave radiation (negative means outgoing)

Rsn : Surface downward net shortwave radiation

Rsi : Surface downward shortwave radiation (Rsi)

Rsi_o : Clear-sky surface downward shortwave radiation

Rsi_toa : Extraterrestrial radiation (top of atmosphere, Rsi_toa)

Note

Rn might <= 0. Users need to constrain the min value by pmax(Rn, 0).

See also

Author

Xie YuXuan and Kong Dongdong

Examples

cal_Rn(lat = 30, J = 1, RH = 70, Tmin = 20, Tmax = 30, ssd = 10)
#>          Rn      Rsn      Rln      Rsi    Rsi_o  Rsi_toa
#> 1: 6.557457 11.45757 -4.90011 14.87996 14.98944 19.98592